terraform/functions/iottotwins/requirements.txt
astroid==2.14.1
azure-core==1.26.3
azure-digitaltwins-core==1.2.0
azure-functions==1.12.0
azure-identity==1.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
dill==0.3.6
idna==3.4
isodate==0.6.1
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
platformdirs==2.6.2
portalocker==2.7.0
pycparser==2.21
PyJWT==2.6.0
pylint==2.16.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.14
wrapt==1.14.1
